Seperating functioonal tests with unit tests
